Fix #5102, translator made more safe
